refactor form automation
